removed state-saving routine (no speed adv)

+++ mfl/compare.c
@@ -118,7 +118,7 @@ tree *mfl_decompress_tree(int *savedtr, int ntax, int numnodes)
     return t;
 }
 
-bool mfl_compare_alltrees(tree *newtopol, tree **savedtrees, int ntax, int numnodes, long int *start, long int *last)
+bool mfl_compare_alltrees(tree *newtopol, tree **savedtrees, int ntax, int numnodes, int *start, long int *last)
 {
     long int i = 0;
     int *newtr;
